Make table structure lock not being held when doing learner read (#864)
Abstract learner read routine, and do learner read before acquiring table's structure lock.
