+++ dbms/src/Flash/Coprocessor/DAGQueryBlockInterpreter.cpp
@@ -25,6 +25,7 @@
 #include <Storages/StorageMergeTree.h>
 #include <Storages/Transaction/CHTableHandle.h>
 #include <Storages/Transaction/KVStore.h>
+#include <Storages/Transaction/LearnerRead.h>
 #include <Storages/Transaction/LockException.h>
 #include <Storages/Transaction/Region.h>
 #include <Storages/Transaction/RegionException.h>
@@ -310,6 +311,62 @@ void DAGQueryBlockInterpreter::executeTS(const tipb::TableScan & ts, Pipeline & 
     const Settings & settings = context.getSettingsRef();
     auto & tmt = context.getTMTContext();
 
+    auto mvcc_query_info = std::make_unique<MvccQueryInfo>();
+    mvcc_query_info->resolve_locks = true;
+    mvcc_query_info->read_tso = settings.read_tso;
+    // We need to validate regions snapshot after getting streams from storage.
+    LearnerReadSnapshot learner_read_snapshot;
+    std::unordered_map<RegionID, const RegionInfo &> region_retry;
+    if (!dag.isBatchCop())
+    {
+        if (auto [info_retry, status] = MakeRegionQueryInfos(dag.getRegions(), {}, tmt, *mvcc_query_info, table_id); info_retry)
+            throw RegionException({(*info_retry).begin()->first}, status);
+
+        learner_read_snapshot = doLearnerRead(table_id, *mvcc_query_info, tmt, log);
+    }
+    else
+    {
+        std::unordered_set<RegionID> force_retry;
+        for (;;)
+        {
+            try
+            {
+                region_retry.clear();
+                auto [retry, status] = MakeRegionQueryInfos(dag.getRegions(), force_retry, tmt, *mvcc_query_info, table_id);
+                std::ignore = status;
+                if (retry)
+                {
+                    region_retry = std::move(*retry);
+                    for (auto & r : region_retry)
+                        force_retry.emplace(r.first);
+                }
+                if (mvcc_query_info->regions_query_info.empty())
+                    break;
+                learner_read_snapshot = doLearnerRead(table_id, *mvcc_query_info, tmt, log);
+                break;
+            }
+            catch (const LockException & e)
+            {
+                // We can also use current thread to resolve lock, but it will block next process.
+                // So, force this region retry in another thread in CoprocessorBlockInputStream.
+                force_retry.emplace(e.region_id);
+            }
+            catch (const RegionException & e)
+            {
+                if (tmt.getTerminated())
+                    throw Exception("TiFlash server is terminating", ErrorCodes::LOGICAL_ERROR);
+                // By now, RegionException will contain all region id of MvccQueryInfo, which is needed by CHSpark.
+                // When meeting RegionException, we can let MakeRegionQueryInfos to check in next loop.
+            }
+            catch (DB::Exception & e)
+            {
+                e.addMessage("(while creating InputStreams from storage `" + storage->getDatabaseName() + "`.`" + storage->getTableName()
+                    + "`, table_id: " + DB::toString(table_id) + ")");
+                throw;
+            }
+        }
+    }
+
     if (settings.schema_version == DEFAULT_UNSPECIFIED_SCHEMA_VERSION)
     {
         storage = context.getTMTContext().getStorages().get(table_id);
@@ -402,9 +459,7 @@ void DAGQueryBlockInterpreter::executeTS(const tipb::TableScan & ts, Pipeline & 
     /// to avoid null point exception
     query_info.query = dummy_query;
     query_info.dag_query = std::make_unique<DAGQueryInfo>(conditions, analyzer->getPreparedSets(), analyzer->getCurrentInputColumns());
-    query_info.mvcc_query_info = std::make_unique<MvccQueryInfo>();
-    query_info.mvcc_query_info->resolve_locks = true;
-    query_info.mvcc_query_info->read_tso = settings.read_tso;
+    query_info.mvcc_query_info = std::move(mvcc_query_info);
 
     if (dag.getRegions().empty())
     {
@@ -419,6 +474,8 @@ void DAGQueryBlockInterpreter::executeTS(const tipb::TableScan & ts, Pipeline & 
         try
         {
             pipeline.streams = storage->read(required_columns, query_info, context, from_stage, max_block_size, max_streams);
+            // After getting streams from storage, we need to validate if regions have changed after learner read.
+            validateQueryInfo(*query_info.mvcc_query_info, learner_read_snapshot, tmt, log);
         }
         catch (DB::Exception & e)
         {
@@ -429,48 +486,25 @@ void DAGQueryBlockInterpreter::executeTS(const tipb::TableScan & ts, Pipeline & 
     }
     else
     {
-        std::unordered_map<RegionID, const RegionInfo &> region_retry;
-        std::unordered_set<RegionID> force_retry;
-        for (;;)
+        // TODO: Note that if storage is (Txn)MergeTree, and any region exception thrown, we won't do retry here.
+        // Now we only support DeltaTree in production environment and don't do any extra check for storage type here.
+        try
         {
-            try
-            {
-                region_retry.clear();
-                auto [retry, status] = MakeRegionQueryInfos(dag.getRegions(), force_retry, tmt, *query_info.mvcc_query_info, table_id);
-                std::ignore = status;
-                if (retry)
-                {
-                    region_retry = std::move(*retry);
-                    for (auto & r : region_retry)
-                        force_retry.emplace(r.first);
-                }
-                if (query_info.mvcc_query_info->regions_query_info.empty())
-                    break;
-                pipeline.streams = storage->read(required_columns, query_info, context, from_stage, max_block_size, max_streams);
-                break;
-            }
-            catch (const LockException & e)
-            {
-                // We can also use current thread to resolve lock, but it will block next process.
-                // So, force this region retry in another thread in CoprocessorBlockInputStream.
-                force_retry.emplace(e.region_id);
-            }
-            catch (const RegionException & e)
-            {
-                if (tmt.getTerminated())
-                    throw Exception("TiFlash server is terminating", ErrorCodes::LOGICAL_ERROR);
-                // By now, RegionException will contain all region id of MvccQueryInfo, which is needed by CHSpark.
-                // When meeting RegionException, we can let MakeRegionQueryInfos to check in next loop.
-            }
-            catch (DB::Exception & e)
-            {
-                e.addMessage("(while creating InputStreams from storage `" + storage->getDatabaseName() + "`.`" + storage->getTableName()
-                    + "`, table_id: " + DB::toString(table_id) + ")");
-                throw;
-            }
+            pipeline.streams = storage->read(required_columns, query_info, context, from_stage, max_block_size, max_streams);
+            // After getting streams from storage, we need to validate whether regions have changed or not after learner read.
+            // In case the versions of regions have changed, those `streams` may contain different data other than expected.
+            // Like after region merge/split.
+            validateQueryInfo(*query_info.mvcc_query_info, learner_read_snapshot, tmt, log);
+        }
+        catch (DB::Exception & e)
+        {
+            e.addMessage("(while creating InputStreams from storage `" + storage->getDatabaseName() + "`.`" + storage->getTableName()
+                + "`, table_id: " + DB::toString(table_id) + ")");
+            throw;
         }
 
-        if (region_retry.size())
+        // For those regions which are not presented in this tiflash node, we will try to fetch streams from other tiflash nodes.
+        if (!region_retry.empty())
         {
             LOG_DEBUG(log, ({
                 std::stringstream ss;
@@ -997,13 +1031,13 @@ SortDescription DAGQueryBlockInterpreter::getSortDescription(std::vector<NameAnd
     order_descr.reserve(topn.order_by_size());
     for (int i = 0; i < topn.order_by_size(); i++)
     {
-	String name = order_columns[i].name;
+        const auto & name = order_columns[i].name;
         int direction = topn.order_by(i).desc() ? -1 : 1;
         // MySQL/TiDB treats NULL as "minimum".
         int nulls_direction = -1;
-	std::shared_ptr<ICollator> collator = nullptr;
-	if (removeNullable(order_columns[i].type)->isString())
-	    collator = getCollatorFromExpr(topn.order_by(i).expr());
+        std::shared_ptr<ICollator> collator = nullptr;
+        if (removeNullable(order_columns[i].type)->isString())
+            collator = getCollatorFromExpr(topn.order_by(i).expr());
 
         order_descr.emplace_back(name, direction, nulls_direction, collator);
     }
