+++ dbms/src/Flash/Coprocessor/InterpreterDAGHelper.hpp
@@ -24,9 +24,9 @@ RegionException::RegionReadStatus GetRegionReadStatus(
     return RegionException::OK;
 }
 
-std::tuple<std::optional<std::unordered_map<RegionID, const RegionInfo &>>, RegionException::RegionReadStatus> MakeRegionQueryInfos(
-    const std::unordered_map<RegionID, RegionInfo> & dag_region_infos, const std::unordered_set<RegionID> & region_force_retry,
-    TMTContext & tmt, MvccQueryInfo & mvcc_info, TableID table_id)
+std::tuple<std::optional<std::unordered_map<RegionID, const RegionInfo &>>, RegionException::RegionReadStatus> //
+MakeRegionQueryInfos(const std::unordered_map<RegionID, RegionInfo> & dag_region_infos,
+    const std::unordered_set<RegionID> & region_force_retry, TMTContext & tmt, MvccQueryInfo & mvcc_info, TableID table_id)
 {
     mvcc_info.regions_query_info.clear();
     std::unordered_map<RegionID, const RegionInfo &> region_need_retry;
