+++ dbms/src/Storages/StorageDeltaMerge.cpp
@@ -24,7 +24,6 @@
 #include <Storages/StorageDeltaMergeHelpers.h>
 #include <Storages/Transaction/KVStore.h>
 #include <Storages/Transaction/Region.h>
-#include <Storages/Transaction/RegionException.h>
 #include <Storages/Transaction/SchemaNameMapper.h>
 #include <Storages/Transaction/TMTContext.h>
 #include <Storages/Transaction/TypeMapping.h>
@@ -316,240 +315,6 @@ void StorageDeltaMerge::write(Block && block, const Settings & settings)
     store->write(global_context, settings, block);
 }
 
-namespace
-{
-
-void throwRetryRegion(const MvccQueryInfo::RegionsQueryInfo & regions_info, RegionException::RegionReadStatus status)
-{
-    std::vector<RegionID> region_ids;
-    region_ids.reserve(regions_info.size());
-    for (const auto & info : regions_info)
-        region_ids.push_back(info.region_id);
-    throw RegionException(std::move(region_ids), status);
-}
-
-/// Check if region is invalid.
-RegionException::RegionReadStatus isValidRegion(const RegionQueryInfo & region_to_query, const RegionPtr & region_in_mem)
-{
-    if (region_in_mem->peerState() != raft_serverpb::PeerState::Normal)
-        return RegionException::RegionReadStatus::NOT_FOUND;
-
-    const auto & [version, conf_ver, key_range] = region_in_mem->dumpVersionRange();
-    (void)key_range;
-    if (version != region_to_query.version || conf_ver != region_to_query.conf_version)
-        return RegionException::RegionReadStatus::VERSION_ERROR;
-
-    return RegionException::RegionReadStatus::OK;
-}
-
-struct RegionMapNode : RegionPtr
-{
-    UInt64 snapshot_event_flag{0};
-
-    RegionMapNode() = default;
-    RegionMapNode(const RegionPtr & region) : RegionPtr(region), snapshot_event_flag(region->getSnapshotEventFlag()) {}
-    bool operator!=(const RegionPtr & tar) const { return (tar != *this) || (tar && snapshot_event_flag != tar->getSnapshotEventFlag()); }
-};
-
-using InternalRegionMap = std::unordered_map<RegionID, RegionMapNode>;
-
-InternalRegionMap doLearnerRead(const TiDB::TableID table_id,   //
-    const MvccQueryInfo::RegionsQueryInfo & regions_query_info, //
-    const bool resolve_locks, const Timestamp start_ts,
-    size_t concurrent_num, //
-    TMTContext & tmt, Poco::Logger * log)
-{
-    assert(log != nullptr);
-
-    MvccQueryInfo::RegionsQueryInfo regions_info;
-    if (!regions_query_info.empty())
-    {
-        regions_info = regions_query_info;
-    }
-    else
-    {
-        // Only for test, because regions_query_info should never be empty if query is from TiDB or TiSpark.
-        auto regions = tmt.getRegionTable().getRegionsByTable(table_id);
-        regions_info.reserve(regions.size());
-        for (const auto & [id, region] : regions)
-        {
-            if (region == nullptr)
-                continue;
-            regions_info.emplace_back(
-                RegionQueryInfo{id, region->version(), region->confVer(), region->getHandleRangeByTable(table_id), {}});
-        }
-    }
-
-    // adjust concurrency by num of regions
-    concurrent_num = std::max(1, std::min(concurrent_num, regions_info.size()));
-
-    KVStorePtr & kvstore = tmt.getKVStore();
-    Context & context = tmt.getContext();
-    InternalRegionMap kvstore_region;
-    // check region is not null and store region map.
-    for (const auto & info : regions_info)
-    {
-        auto region = kvstore->getRegion(info.region_id);
-        if (region == nullptr)
-        {
-            LOG_WARNING(log, "[region " << info.region_id << "] is not found in KVStore, try again");
-            throwRetryRegion(regions_info, RegionException::RegionReadStatus::NOT_FOUND);
-        }
-        kvstore_region.emplace(info.region_id, std::move(region));
-    }
-    // make sure regions are not duplicated.
-    if (unlikely(kvstore_region.size() != regions_info.size()))
-        throw Exception("Duplicate region id", ErrorCodes::LOGICAL_ERROR);
-
-    auto metrics = context.getTiFlashMetrics();
-    const size_t num_regions = regions_info.size();
-    const size_t batch_size = num_regions / concurrent_num;
-    std::atomic_uint8_t region_status = RegionException::RegionReadStatus::OK;
-    const auto batch_wait_index = [&, resolve_locks, start_ts](const size_t region_begin_idx) -> void {
-        const size_t region_end_idx = std::min(region_begin_idx + batch_size, num_regions);
-        for (size_t region_idx = region_begin_idx; region_idx < region_end_idx; ++region_idx)
-        {
-            // If any threads meets an error, just return.
-            if (region_status != RegionException::RegionReadStatus::OK)
-                return;
-
-            RegionQueryInfo & region_to_query = regions_info[region_idx];
-            const RegionID region_id = region_to_query.region_id;
-            auto region = kvstore_region[region_id];
-
-            auto status = isValidRegion(region_to_query, region);
-            if (status != RegionException::RegionReadStatus::OK)
-            {
-                region_status = status;
-                LOG_WARNING(log,
-                    "Check memory cache, region " << region_id << ", version " << region_to_query.version << ", handle range ["
-                                                  << region_to_query.range_in_table.first.toString() << ", "
-                                                  << region_to_query.range_in_table.second.toString() << ") , status "
-                                                  << RegionException::RegionReadStatusString(status));
-                return;
-            }
-
-            GET_METRIC(metrics, tiflash_raft_read_index_count).Increment();
-            Stopwatch read_index_watch;
-
-            /// Blocking learner read. Note that learner read must be performed ahead of data read,
-            /// otherwise the desired index will be blocked by the lock of data read.
-            auto read_index_result = region->learnerRead();
-            GET_METRIC(metrics, tiflash_raft_read_index_duration_seconds).Observe(read_index_watch.elapsedSeconds());
-            if (read_index_result.region_unavailable)
-            {
-                // client-c detect region removed. Set region_status and continue.
-                region_status = RegionException::RegionReadStatus::NOT_FOUND;
-                continue;
-            }
-            else if (read_index_result.region_epoch_not_match)
-            {
-                region_status = RegionException::RegionReadStatus::VERSION_ERROR;
-                continue;
-            }
-            else
-            {
-                Stopwatch wait_index_watch;
-                if (region->waitIndex(read_index_result.read_index, tmt.getTerminated()))
-                {
-                    region_status = RegionException::RegionReadStatus::NOT_FOUND;
-                    continue;
-                }
-                GET_METRIC(metrics, tiflash_raft_wait_index_duration_seconds).Observe(wait_index_watch.elapsedSeconds());
-            }
-            if (resolve_locks)
-            {
-                status = RegionTable::resolveLocksAndWriteRegion( //
-                    tmt,                                          //
-                    table_id,                                     //
-                    region,                                       //
-                    start_ts,                                     //
-                    region_to_query.bypass_lock_ts,               //
-                    region_to_query.version,                      //
-                    region_to_query.conf_version,                 //
-                    region_to_query.range_in_table, log);
-
-                if (status != RegionException::RegionReadStatus::OK)
-                {
-                    LOG_WARNING(log,
-                        "Check memory cache, region " << region_id << ", version " << region_to_query.version << ", handle range ["
-                                                      << region_to_query.range_in_table.first.toString() << ", "
-                                                      << region_to_query.range_in_table.second.toString() << ") , status "
-                                                      << RegionException::RegionReadStatusString(status));
-                    region_status = status;
-                }
-            }
-        }
-    };
-    auto start_time = Clock::now();
-    if (concurrent_num <= 1)
-    {
-        batch_wait_index(0);
-    }
-    else
-    {
-        ::ThreadPool pool(concurrent_num);
-        for (size_t region_begin_idx = 0; region_begin_idx < num_regions; region_begin_idx += batch_size)
-        {
-            pool.schedule([&batch_wait_index, region_begin_idx] { batch_wait_index(region_begin_idx); });
-        }
-        pool.wait();
-    }
-
-    // Check if any region is invalid, TiDB / TiSpark should refresh region cache and retry.
-    if (region_status != RegionException::RegionReadStatus::OK)
-        throwRetryRegion(regions_info, static_cast<RegionException::RegionReadStatus>(region_status.load()));
-
-    auto end_time = Clock::now();
-    LOG_DEBUG(log,
-        "[Learner Read] wait index cost " << std::chrono::duration_cast<std::chrono::milliseconds>(end_time - start_time).count() << " ms");
-
-    /// If background flush is disabled, we don't need to do both flushing KVStore or waiting for background tasks.
-    if (!tmt.isBgFlushDisabled())
-    {
-        // After raft index is satisfied, we flush region to StorageDeltaMerge so that we can read all data
-        start_time = Clock::now();
-        std::set<RegionID> regions_flushing_in_bg_threads;
-        auto & region_table = tmt.getRegionTable();
-        for (auto && [region_id, region] : kvstore_region)
-        {
-            if (!region->dataSize())
-                continue;
-            auto to_remove_data = region_table.tryFlushRegion(region, false);
-            // If region is flushing by other bg threads, we should mark those regions to wait.
-            if (to_remove_data.empty())
-            {
-                regions_flushing_in_bg_threads.insert(region_id);
-                LOG_DEBUG(log, "[Learner Read] region " << region_id << " is flushing by other thread.");
-            }
-        }
-        end_time = Clock::now();
-        LOG_DEBUG(log,
-            "[Learner Read] flush " << kvstore_region.size() - regions_flushing_in_bg_threads.size() << " regions of "
-                                    << kvstore_region.size() << " to StorageDeltaMerge cost "
-                                    << std::chrono::duration_cast<std::chrono::milliseconds>(end_time - start_time).count() << " ms");
-
-        // Maybe there is some data not flush to store yet, we should wait till all regions is flushed.
-        if (!regions_flushing_in_bg_threads.empty())
-        {
-            start_time = Clock::now();
-            for (const auto & region_id : regions_flushing_in_bg_threads)
-            {
-                region_table.waitTillRegionFlushed(region_id);
-            }
-            end_time = Clock::now();
-            LOG_DEBUG(log,
-                "[Learner Read] wait bg flush " << regions_flushing_in_bg_threads.size() << " regions to StorageDeltaMerge cost "
-                                                << std::chrono::duration_cast<std::chrono::milliseconds>(end_time - start_time).count()
-                                                << " ms");
-        }
-    }
-
-    return kvstore_region;
-}
-
-} // namespace
-
 std::unordered_set<UInt64> parseSegmentSet(const ASTPtr & ast)
 {
     if (!ast)
@@ -678,20 +443,10 @@ BlockInputStreams StorageDeltaMerge::read( //
         /// Else a request comes from CH-client, we set `concurrent_num` by num_streams.
         size_t concurrent_num = std::max<size_t>(num_streams * mvcc_query_info.concurrent, 1);
 
-        // With `no_kvstore` is true, we do not do learner read
-        InternalRegionMap regions_in_learner_read;
-        if (likely(!select_query.no_kvstore))
-        {
-            /// Learner read.
-            regions_in_learner_read = doLearnerRead(tidb_table_info.id, mvcc_query_info.regions_query_info, mvcc_query_info.resolve_locks,
-                mvcc_query_info.read_tso, concurrent_num, tmt, log);
-
-            if (likely(!mvcc_query_info.regions_query_info.empty()))
-            {
-                /// For learner read from TiDB/TiSpark, we set num_streams by `concurrent_num`
-                num_streams = concurrent_num;
-            } // else learner read from ch-client, keep num_streams
-        }
+        /// For learner read from TiDB/TiSpark, we set num_streams by `concurrent_num`
+        if (likely(!select_query.no_kvstore && !mvcc_query_info.regions_query_info.empty()))
+            num_streams = concurrent_num;
+        // else learner read from ch-client, keep num_streams
 
         HandleRanges ranges = getQueryRanges(mvcc_query_info.regions_query_info);
 
@@ -763,37 +518,8 @@ BlockInputStreams StorageDeltaMerge::read( //
         auto streams = store->read(context, context.getSettingsRef(), columns_to_read, ranges, num_streams,
             /*max_version=*/mvcc_query_info.read_tso, rs_operator, max_block_size, parseSegmentSet(select_query.segment_expression_list));
 
-        {
-            /// Ensure read_tso and regions' info after read.
-
-            check_read_tso(mvcc_query_info.read_tso);
-
-            for (const auto & region_query_info : mvcc_query_info.regions_query_info)
-            {
-                RegionException::RegionReadStatus status = RegionException::RegionReadStatus::OK;
-                auto region = tmt.getKVStore()->getRegion(region_query_info.region_id);
-                if (regions_in_learner_read[region_query_info.region_id] != region)
-                    status = RegionException::RegionReadStatus::NOT_FOUND;
-                else if (region->version() != region_query_info.version)
-                {
-                    // ABA problem may cause because one region is removed and inserted back.
-                    // if the version of region is changed, the `streams` may has less data because of compaction.
-                    status = RegionException::RegionReadStatus::VERSION_ERROR;
-                }
-
-                if (status != RegionException::RegionReadStatus::OK)
-                {
-                    LOG_WARNING(log,
-                        "Check after read from DeltaMergeStore, region "
-                            << region_query_info.region_id << ", version " << region_query_info.version //
-                            << ", handle range [" << region_query_info.range_in_table.first.toString() << ", "
-                            << region_query_info.range_in_table.second.toString() << "), status "
-                            << RegionException::RegionReadStatusString(status));
-                    // throw region exception and let TiDB retry
-                    throwRetryRegion(mvcc_query_info.regions_query_info, status);
-                }
-            }
-        }
+        /// Ensure read_tso info after read.
+        check_read_tso(mvcc_query_info.read_tso);
 
         return streams;
     }
